backend/requirements.txt
Django>=5.0,<6.0
psycopg[binary]>=3.1
Pillow>=10.0
django-sass-processor>=1.2
libsass>=0.23.0
django-compressor>=4.1
python-dotenv
django-simple-captcha
gunicorn